requirements.txt
opencv-python==4.8.1.78
mediapipe==0.10.8
numpy==1.24.3
tensorflow-macos==2.15.0
tensorflow-metal==1.1.0
matplotlib==3.7.1
scikit-learn==1.3.0